Prefer src for vite app, better package type exports
