+++ excalidraw-app/app-language/LanguageList.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useI18n, languages } from "../../packages/excalidraw/i18n";
+import { useI18n, languages } from "@excalidraw/excalidraw/i18n";
 import { useSetAtom } from "../app-jotai";
 import { appLangCodeAtom } from "./language-state";
