+++ excalidraw-app/tests/LanguageList.test.tsx
@@ -1,11 +1,11 @@
-import { defaultLang } from "../../packages/excalidraw/i18n";
-import { UI } from "../../packages/excalidraw/tests/helpers/ui";
+import { defaultLang } from "@excalidraw/excalidraw/i18n";
+import { UI } from "@excalidraw/excalidraw/tests/helpers/ui";
 import {
   screen,
   fireEvent,
   waitFor,
   render,
-} from "../../packages/excalidraw/tests/test-utils";
+} from "@excalidraw/excalidraw/tests/test-utils";
 
 import ExcalidrawApp from "../App";
