+++ packages/excalidraw/clients.ts
@@ -3,6 +3,7 @@ import {
   COLOR_VOICE_CALL,
   COLOR_WHITE,
   THEME,
+  UserIdleState,
 } from "./constants";
 import { roundRect } from "./renderer/roundRect";
 import type { InteractiveCanvasRenderConfig } from "./scene/types";
@@ -11,7 +12,6 @@ import type {
   InteractiveCanvasAppState,
   SocketId,
 } from "./types";
-import { UserIdleState } from "./types";
 
 function hashToInteger(id: string) {
   let hash = 0;
