+++ packages/excalidraw/constants.ts
@@ -461,3 +461,9 @@ export const ELEMENT_LINK_KEY = "element";
 
 /** used in tests */
 export const ORIG_ID = Symbol.for("__test__originalId__");
+
+export enum UserIdleState {
+  ACTIVE = "active",
+  AWAY = "away",
+  IDLE = "idle",
+}
