+++ packages/excalidraw/tests/cropElement.test.tsx
@@ -314,6 +314,7 @@ describe("Cropping and other features", async () => {
 
     const canvas = await exportToCanvas({
       elements: [image],
+      // @ts-ignore
       appState: h.state,
       files: h.app.files,
       exportPadding: 0,
@@ -324,6 +325,7 @@ describe("Cropping and other features", async () => {
 
     const svg = await exportToSvg({
       elements: [image],
+      // @ts-ignore
       appState: h.state,
       files: h.app.files,
       exportPadding: 0,
